moved HTML5JavaScriptInterface into its own class
for https://github.com/FieldDB/FieldDB/issues/2085
