+++ src/com/github/fielddb/javascript/HTML5Activity.java
@@ -219,31 +219,6 @@ public abstract class HTML5Activity extends Activity {
 
   }
 
-  public class HTML5JavaScriptInterface extends JavaScriptInterface {
-    private static final long serialVersionUID = 373085850425945181L;
-
-    HTML5Activity mUIParent;
-
-    public HTML5JavaScriptInterface(String outputDir, Context context, HTML5Activity UIParent,
-        String assetsPrefix) {
-      super(outputDir, context, UIParent, assetsPrefix);
-    }
-
-    public HTML5JavaScriptInterface(Context context) {
-      super(context);
-    }
-
-    @Override
-    public HTML5Activity getUIParent() {
-      return this.mUIParent;
-    }
-
-    @Override
-    public void setUIParent(HTML5Activity UIParent) {
-      this.mUIParent = UIParent;
-    }
-  }
-
   public static Bitmap decodeBase64(String input) {
     byte[] decodedByte = Base64.decode(input, 0);
     return BitmapFactory.decodeByteArray(decodedByte, 0, decodedByte.length);
